tools/z3fdb-profiler/hpc-jobs/requirements.txt
earthkit-data
zarr>=3
